Added clarification
The aim of this small project is to use on pivot boxes as readily alternative to RustScan, GoScan and Amass without the need to install other tools. Granted if you are on a Windows box it would be wiser to use a Powershell implementation rather than this. Think of this script as a last resort to other tools.